src/database/config.py
import os
from dotenv import load_dotenv

load_dotenv()

SUPABASE_URL = os.getenv("SUPABASE_URL")
SUPABASE_KEY = os.getenv("SUPABASE_KEY")
EMAIL_REMITENTE = os.getenv("EMAIL_REMITENTE")
EMAIL_PASSWORD = os.getenv("EMAIL_PASSWORD")
EMAIL_DESTINO = os.getenv("EMAIL_DESTINO")

# Configuración de tablas
TABLES = {
    "usuarios": "usuarios",
    "productos": "productos",
    "despachos": "despachos",
    "calendario_pagos": "calendario_pagos"
}
